requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pop==6.0.0
pylint==2.3.1
python-decouple==3.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
